Lots of work on creating meetings and retrieving transcripts

+++ teamv/views.py
@@ -10,7 +10,7 @@ from chat import ChatNamespace
 from os import environ, path
 from dbconnect import connect_to_db
 
-import os, time, datetime
+import os, time, datetime, logging
 
 mylookup = TemplateLookup(directories=['./teamv/templates'], module_directory='./teamv/temp/mako_modules', collection_size=500)
 
@@ -48,15 +48,30 @@ def index(request):
     
 @view_config(route_name='transcript', renderer='mako')
 def transcript(request):
-    # TODO: Link to CDN instead of local logs
-    if 'meeting' in request.GET and is_number(request.GET.get('meeting')):
-        file_name = 'teamv/temp/logs/log_{0}.log'.format(request.GET.get('meeting'))
-        if os.path.isfile(file_name):
+    this_meeting = request.matchdict['meeting']
+
+    if is_number(this_meeting):
+        conn = connect_to_db()
+        cur = conn.cursor()
+
+        cur.execute('SELECT time_started, time_finished FROM meetings WHERE meeting=%s', (this_meeting, ))
+        result = cur.fetchone()
+
+        cur.close()
+        conn.close()
+
+        file_name = 'teamv/temp/logs/log_{0}.log'.format(this_meeting)
+        if result is None:
+            return not_found(request)
+        elif result[0] is not None and result[1] is None and os.path.isfile(file_name): # Meeting is in process and still in local memory
             mytemplate = mylookup.get_template('transcript.mak')
-            this_meeting = request.GET.get('meeting')
             result = mytemplate.render(title = 'Team Valente - Transcript {0}'.format(this_meeting), meeting = this_meeting)
             return Response(result)
-        else:
+        elif result[0] is not None and result[1] is not None: # Meeting is done, must be in CDN
+            # TODO: If log is in CDN, retrieve it
+            print 'Must get file from CDN. But for now, not found.'
+            return not_found(request)
+        else: # Something went wrong
             return not_found(request)
     else:
         return not_found(request)
@@ -64,31 +79,33 @@ def transcript(request):
 @view_config(route_name='start_meeting')
 def start_meeting(request):
     if 'meeting' in request.POST and is_number(request.POST.get('meeting')):
-        # TODO: Check if file in CDN instead of local
         file_name = 'teamv/temp/logs/log_{0}.log'.format(request.POST.get('meeting'))
         conn = connect_to_db()
         cur = conn.cursor()
 
-        if os.path.isfile(file_name): # If the meeting has already been created, check if the meeting has ended
-            cur.execute('SELECT time_started, time_finished FROM meetings WHERE meeting=%s', (request.POST.get('meeting'), ))
+        cur.execute('SELECT time_started, time_finished FROM meetings WHERE meeting=%s', (request.POST.get('meeting'), ))
+        result = cur.fetchone()
 
-            result = cur.fetchone()
-            cur.close()
-            conn.close()
-        
-            if result is None:
-                return Response(status = '404 Not Found') # Something has gone wrong, meeting not found
-            elif result[1] is not None:
-                return Response(status = '403 Forbidden') # Meeting has ended, forbidden
-            else:
-                return Response(status = '200 OK') # Meeting is still going, it is OK to join
-        else:
+        if result is None and not os.path.isfile(file_name): # Meeting is open but not created, so create it
             open(file_name, 'w').close()
             cur.execute('INSERT INTO meetings (meeting) VALUES (%s)', (request.POST.get('meeting'), ))
-
+            print 'INSERT INTO meetings (meeting) VALUES ({0})'.format(request.POST.get('meeting'))
             cur.close()
             conn.close()
-            return Response(status = '201 Created') # Meeting has not been started yet, so it is created
+            return Response(status = '201 Created')
+        elif result[0] is not None and result[1] is None and os.path.isfile(file_name): # Meeting is still going, it is OK to join
+            cur.close()
+            conn.close()
+            return Response(status = '200 OK')
+        elif result[0] is not None and result[1] is not None: # Meeting has ended, forbidden
+            cur.close()
+            conn.close()
+            return Response(status = '403 Forbidden')
+        else: # Something has gone wrong, meeting not found
+            cur.close()
+            conn.close()
+            logging.error('Error creating meeting')
+            return Response(status = '404 Not Found')
     else: # Invalid request
         return Response(status = '400 Bad Request') # No meeting number specified, bad request
